moved Credentials to separate file
