+++ pkg/docker/config.go
@@ -12,20 +12,6 @@ import (
 	perrors "github.com/pkg/errors"
 )
 
-// Credentials holds the information shared between docker and the credentials store.
-type Credentials struct {
-	ServerURL string
-	Username  string
-	Secret    string
-}
-
-func (c *Credentials) AuthToken() string {
-	if c.Username != "" {
-		return c.Username + ":" + c.Secret
-	}
-	return c.Secret
-}
-
 const dockerFileFolder = ".docker"
 const AzureContainerRegistryUsername = "00000000-0000-0000-0000-000000000000"
